Minor types

+++ src/leg.py
@@ -45,8 +45,8 @@ bond_pairs = jnp.array([
     [-4, -1],
     [-3, -1],
     [-2, -1]
-])
-vertices = jnp.arange(0, 6*12, 6)
+], dtype=jnp.int32)
+vertices = jnp.arange(0, 6*12, 6, dtype=jnp.int32)
 def get_leg_energy_fn(soft_sphere_eps, bond_diameter, shape, shape_species):
     def leg_energy_fn(body):
         position = rigid_body.union_to_points(body, shape, shape_species)
